Login and Employee info systems added
Login system is completed with client-side validations on the form input fields. Employee information display is created in the form of table, also link for page where employee information can be added and will contain delete/redact forms for the employee information. Most of the Back-End is done, almost no work on the Front-End.